requirements.txt
sqlalchemy
numpy
alembic
python-dotenv
greenlet
sqlite-vec
feedparser
requests
lxml
goose3
loguru
grequests
sentence-transformers
transformers
torch
torchaudio
torchvision
bertopic
tenacity
fastapi
uvicorn[standard]
httpx
pytest